+++ frontend/src/components/AccountMenu.vue
@@ -51,6 +51,9 @@ async function logout() {
       <RouterLink class="account-menu-item" to="/account" @click="closeMenu">
         Account settings
       </RouterLink>
+      <RouterLink class="account-menu-item" to="/favorites" @click="closeMenu">
+        Favorite lists
+      </RouterLink>
       <button class="account-menu-item" type="button" @click="logout">Sign out</button>
     </div>
   </details>